// src/services/business/types.ts
export type PostCardLayoutImages = 'static' | 'hoverZoom' | 'slider' | 'switch' | 'rounded';

export type PostCardSize = 'small' | 'medium' | 'long';
export type PostCardLayoutName = 'none' | 'basic';
export type PostCardLayoutPrice = 'none' | 'basic' | 'smallerCurrency' | 'usdCurrencySymbol';
export type PostCardLayoutDiscount = 'none' | 'savedPercent' | 'savedMoney';
export type PostLayoutShoppingMethod = 'none' | 'shoppingCart';

export type PostLayoutShoppingCart = 'none' | 'added_whatsApp';

export type PostCardLayoutMetaLayout = 'basic' | 'verticalCentered';

export interface PostCardLayout {
  images?: PostCardLayoutImages;
  size?: PostCardSize;
  metaLayout?: PostCardLayoutMetaLayout;
  name?: PostCardLayoutName;
  price?: PostCardLayoutPrice;
  discount?: PostCardLayoutDiscount;
  shoppingMethod?: PostLayoutShoppingMethod;
}
// ...
export enum DeliveryConfigType {
  NONE = 'NONE',
  FREE = 'FREE',
  REQUIRED = 'REQUIRED',
  OPTIONAL = 'OPTIONAL'
}

export interface DeliveryConfig {
  minPrice?: number;
  priceByKm?: number;
  type?: DeliveryConfigType;
}